Uploaded 2 new exercises and removed 5 tasks in the "More exercise" folder of the '02JS-Arrays-and-Strings' topic. The tasks were initially from a different topic, folders got mixed up. The current tasks are now aligned with the correct theme.
In this commit, I addressed a folder mix-up issue in the "More exercise" directory of the '02JS-Arrays-and-Strings' topic. Initially, I had uploaded 5 tasks that belonged to a different topic, causing confusion. To resolve this, I have removed those tasks and replaced them with 2 new exercises that are now in line with the intended theme. This ensures better organization and clarity within the repository.